refactor(Cars) : Car 객체들의 상태를 관리해줄 Cars 객체 생성
1. Racing의 책임 분리, List<Car>에서 하던 검증 및 출력 역할을 Cars객체로 이관
2. Race 시 임의 번호를 생성하는 책임을 Cars로 옮김으로서 Car의 Race를 테스트 할 수 있게 수정
