+++ src/main/java/racingcar/Car.java
@@ -1,10 +1,6 @@
 package racingcar;
 
-import utils.RandomUtils;
-
 public class Car {
-    public static final int MIN_VALUE = 0;
-    public static final int MAX_VALUE = 9;
     public static final int STOP_NUMBER = 3;
     private final String name;
     private int position = 0;
@@ -13,8 +9,8 @@ public class Car {
         this.name = name;
     }
 
-    public void race() {
-        if (RandomUtils.nextInt(MIN_VALUE, MAX_VALUE) > STOP_NUMBER) {
+    public void race(int randomNumber) {
+        if (randomNumber > STOP_NUMBER) {
             position++;
         }
     }
