+++ src/main/java/racingcar/Racing.java
@@ -1,16 +1,14 @@
 package racingcar;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class Racing {
     public static final String COMMA_AND_SPACE = ", ";
     public static final String FINAL_WINNER_MSG = "최종 우승자 : ";
-    public static final String COLON_WRAPPED_BLANK = " : ";
-    private List<Car> cars;
+    private Cars cars;
     private int round;
 
-    public Racing(List<Car> cars, int round) {
+    public Racing(Cars cars, int round) {
         this.cars = cars;
         this.round = round;
     }
@@ -20,42 +18,17 @@ public class Racing {
     }
 
     public void nextRound() {
-        for (Car car : cars) {
-            car.race();
-        }
+        cars.race();
         round--;
     }
 
     public String printProgress() {
-        StringBuffer progress = new StringBuffer();
-        for (Car car : cars) {
-            progress.append(car.getName()).append(COLON_WRAPPED_BLANK).append(revertPositionToDash(car.getPosition())).append(System.lineSeparator());
-        }
-        return progress.toString();
+        return cars.printProgress();
     }
 
     public String printResult() {
-        int maxPosition = 0;
-        for (Car car : cars) {
-            if (car.getPosition() > maxPosition) {
-                maxPosition = car.getPosition();
-            }
-        }
-        List<String> winners = new ArrayList<>();
-        for (Car car : cars) {
-            if (car.getPosition() == maxPosition) {
-                winners.add(car.getName());
-            }
-        }
-
+        List<String> winners = cars.getWinners();
         return FINAL_WINNER_MSG + String.join(COMMA_AND_SPACE, winners);
     }
 
-    private String revertPositionToDash(int position) {
-        StringBuffer dash = new StringBuffer();
-        for (int i = 0; i < position; i++) {
-            dash.append("-");
-        }
-        return dash.toString();
-    }
 }
